requirements/base.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic-settings==2.12.0
python-dotenv==1.2.1
requests==2.32.3
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-json-logger==3.3.0
redis==5.0.0
